prices: wNEAR→NEAR alias + CoinGecko historical fallback (NPRO, stNEAR, …)
History previously came only from CryptoCompare (majors), so NEAR-ecosystem
tokens had no price history and the report errored/prompted on them.

- token-map.js: extract the symbol/coingecko-id map + add SYMBOL_ALIASES
  (wnear→near, usdt.e→usdt, usdc.e→usdc) so wrapped/bridged tokens price as
  their canonical asset (NEAR has long CryptoCompare history).
- getDailyPrice/index: fall back to CoinGecko market_chart (by coin id, last
  365d) when CryptoCompare doesn't list a symbol — covers NPRO, stNEAR, SHITZU,
  etc. EOD updater uses the same fallback.
- Unknown/scam tokens now resolve to empty history (graceful) instead of a 500.

Verified live: wNEAR/USDT.e/stNEAR/NPRO/SHITZU all return history. 38 tests pass.
